+++ framework/application/application.h
@@ -87,10 +87,7 @@ class Application final
 
     void StopRunning() { running_ = false; }
 
-    uint32_t GetCurrentFrameNumber() const
-    {
-        return file_processor_->GetCurrentFrameNumber();
-    }
+    uint32_t GetCurrentFrameNumber() const { return static_cast<uint32_t>(file_processor_->GetCurrentFrameNumber()); }
 
   private:
     // clang-format off
